Refactor flutter-winit.
