+++ flutter-winit/src/window.rs
@@ -1,10 +1,8 @@
 use crate::context::Context;
 use crate::handler::{WinitFlutterEngineHandler, WinitPlatformHandler, WinitWindowHandler};
+use crate::keyboard::raw_key;
+use crate::pointer::Pointers;
 use flutter_engine::channel::Channel;
-use flutter_engine::ffi::{
-    FlutterPointerDeviceKind, FlutterPointerMouseButtons, FlutterPointerPhase,
-    FlutterPointerSignalKind,
-};
 use flutter_engine::plugins::Plugin;
 use flutter_engine::texture_registry::TextureRegistry;
 use flutter_engine::{FlutterEngine, FlutterEngineHandler};
@@ -20,8 +18,7 @@ use flutter_plugins::system::SystemPlugin;
 use flutter_plugins::textinput::TextInputPlugin;
 use flutter_plugins::window::WindowPlugin;
 use glutin::event::{
-    DeviceId, ElementState, Event, KeyboardInput, MouseButton, MouseScrollDelta, Touch, TouchPhase,
-    VirtualKeyCode, WindowEvent,
+    ElementState, Event, KeyboardInput, MouseScrollDelta, Touch, VirtualKeyCode, WindowEvent,
 };
 use glutin::event_loop::{ControlFlow, EventLoop};
 use glutin::window::WindowBuilder;
@@ -359,275 +356,4 @@ fn resize(engine: &FlutterEngine, context: &Arc<Mutex<Context>>) {
     );
     context.resize(size);
     engine.send_window_metrics_event(size.width as usize, size.height as usize, dpi);
-}
-
-struct Pointers {
-    engine: FlutterEngine,
-    pointers: Vec<Pointer>,
-}
-
-struct Pointer {
-    device_id: DeviceId,
-    touch: bool,
-    position: (f64, f64),
-    pressed: u32,
-}
-
-impl Pointer {
-    fn new(device_id: DeviceId, touch: bool) -> Self {
-        Self {
-            device_id,
-            touch,
-            position: (0.0, 0.0),
-            pressed: 0,
-        }
-    }
-}
-
-impl Pointers {
-    fn new(engine: FlutterEngine) -> Self {
-        Self {
-            engine,
-            pointers: Default::default(),
-        }
-    }
-
-    fn index(&mut self, device_id: DeviceId, touch: bool) -> usize {
-        if let Some(index) = self
-            .pointers
-            .iter()
-            .position(|p| p.device_id == device_id && p.touch == touch)
-        {
-            index
-        } else {
-            let index = self.pointers.len();
-            self.pointers.push(Pointer::new(device_id, touch));
-            index
-        }
-    }
-
-    fn enter(&mut self, device_id: DeviceId) {
-        let device = self.index(device_id, false);
-        let pointer = &self.pointers[device];
-        self.engine.send_pointer_event(
-            device as i32 + 10,
-            FlutterPointerPhase::Add,
-            pointer.position,
-            FlutterPointerSignalKind::None,
-            (0.0, 0.0),
-            FlutterPointerDeviceKind::Mouse,
-            FlutterPointerMouseButtons::Primary,
-        );
-    }
-
-    fn leave(&mut self, device_id: DeviceId) {
-        let device = self.index(device_id, false);
-        let pointer = &self.pointers[device];
-        self.engine.send_pointer_event(
-            device as i32 + 10,
-            FlutterPointerPhase::Remove,
-            pointer.position,
-            FlutterPointerSignalKind::None,
-            (0.0, 0.0),
-            FlutterPointerDeviceKind::Mouse,
-            FlutterPointerMouseButtons::Primary,
-        );
-    }
-
-    fn moved(&mut self, device_id: DeviceId, position: (f64, f64)) {
-        let device = self.index(device_id, false);
-        self.pointers[device].position = position;
-        let pointer = &self.pointers[device];
-        let phase = if pointer.pressed == 0 {
-            FlutterPointerPhase::Hover
-        } else {
-            FlutterPointerPhase::Move
-        };
-        self.engine.send_pointer_event(
-            device as i32 + 10,
-            phase,
-            pointer.position,
-            FlutterPointerSignalKind::None,
-            (0.0, 0.0),
-            FlutterPointerDeviceKind::Mouse,
-            FlutterPointerMouseButtons::Primary,
-        );
-    }
-
-    fn input(&mut self, device_id: DeviceId, state: ElementState, button: MouseButton) {
-        let device = self.index(device_id, false);
-        match state {
-            ElementState::Pressed => self.pointers[device].pressed += 1,
-            ElementState::Released => self.pointers[device].pressed -= 1,
-        }
-        let pointer = &self.pointers[device];
-        let phase = match state {
-            ElementState::Pressed => FlutterPointerPhase::Down,
-            ElementState::Released => FlutterPointerPhase::Up,
-        };
-        let button = match button {
-            MouseButton::Left => FlutterPointerMouseButtons::Primary,
-            MouseButton::Right => FlutterPointerMouseButtons::Secondary,
-            MouseButton::Middle => FlutterPointerMouseButtons::Middle,
-            MouseButton::Other(4) => FlutterPointerMouseButtons::Back,
-            MouseButton::Other(5) => FlutterPointerMouseButtons::Forward,
-            _ => FlutterPointerMouseButtons::Primary,
-        };
-        self.engine.send_pointer_event(
-            device as i32 + 10,
-            phase,
-            pointer.position,
-            FlutterPointerSignalKind::None,
-            (0.0, 0.0),
-            FlutterPointerDeviceKind::Mouse,
-            button,
-        );
-    }
-
-    fn wheel(&mut self, device_id: DeviceId, delta: (f64, f64)) {
-        let device = self.index(device_id, false);
-        let pointer = &self.pointers[device];
-        let phase = if pointer.pressed == 0 {
-            FlutterPointerPhase::Hover
-        } else {
-            FlutterPointerPhase::Move
-        };
-        self.engine.send_pointer_event(
-            device as i32 + 10,
-            phase,
-            pointer.position,
-            FlutterPointerSignalKind::Scroll,
-            delta,
-            FlutterPointerDeviceKind::Mouse,
-            FlutterPointerMouseButtons::Primary,
-        );
-    }
-
-    fn touch(&mut self, device_id: DeviceId, phase: TouchPhase, position: (f64, f64)) {
-        let device = self.index(device_id, true);
-        let phase = match phase {
-            TouchPhase::Started => {
-                self.pointers[device].pressed += 1;
-                FlutterPointerPhase::Down
-            }
-            TouchPhase::Moved => FlutterPointerPhase::Move,
-            TouchPhase::Ended => FlutterPointerPhase::Up,
-            TouchPhase::Cancelled => FlutterPointerPhase::Cancel,
-        };
-        self.engine.send_pointer_event(
-            self.pointers[device].pressed as i32 - 1,
-            phase,
-            position,
-            FlutterPointerSignalKind::None,
-            (0.0, 0.0),
-            FlutterPointerDeviceKind::Touch,
-            FlutterPointerMouseButtons::Primary,
-        );
-        if phase == FlutterPointerPhase::Up || phase == FlutterPointerPhase::Cancel {
-            self.pointers[device].pressed -= 1;
-        }
-    }
-}
-
-// Emulates glfw key numbers
-// https://github.com/flutter/flutter/blob/master/packages/flutter/lib/src/services/keyboard_maps.dart
-fn raw_key(key: Option<VirtualKeyCode>) -> Option<u32> {
-    let key = if let Some(key) = key {
-        if key as u32 >= Key::A as u32 && key as u32 <= Key::Z as u32 {
-            return Some(key as u32 - Key::A as u32 + 65);
-        }
-
-        if key as u32 >= Key::Key1 as u32 && key as u32 <= Key::Key9 as u32 {
-            return Some(key as u32 - Key::Key1 as u32 + 49);
-        }
-
-        key
-    } else {
-        return None;
-    };
-
-    use VirtualKeyCode as Key;
-    let code = match key {
-        Key::Key0 => 48,
-        Key::Return => 257,
-        Key::Escape => 256,
-        Key::Back => 259,
-        Key::Tab => 258,
-        Key::Space => 32,
-        Key::LControl => 341,
-        Key::LShift => 340,
-        Key::LAlt => 342,
-        Key::LWin => 343,
-        Key::RControl => 345,
-        Key::RShift => 344,
-        Key::RAlt => 346,
-        Key::RWin => 347,
-        Key::Minus => 45,
-        Key::Equals => 61,
-        Key::LBracket => 91,
-        Key::RBracket => 93,
-        Key::Backslash => 92,
-        Key::Semicolon => 59,
-        Key::Apostrophe => 39,
-        //Key::Backquote => 96,
-        Key::Comma => 44,
-        Key::Period => 46,
-        Key::Slash => 47,
-        //Key::CapsLock => 280,
-        Key::Snapshot => 283,
-        Key::Pause => 284,
-        Key::Insert => 260,
-        Key::Home => 268,
-        Key::PageUp => 266,
-        Key::Delete => 261,
-        Key::End => 269,
-        Key::PageDown => 267,
-        Key::Right => 262,
-        Key::Left => 263,
-        Key::Down => 264,
-        Key::Up => 265,
-        Key::Numlock => 282,
-        //Key::NumpadDivide => 331,
-        //Key::NumpadMultiply => 332,
-        //Key::NumpadAdd => 334,
-        Key::NumpadEnter => 335,
-        Key::Numpad0 => 320,
-        Key::Numpad1 => 321,
-        Key::Numpad2 => 322,
-        Key::Numpad3 => 323,
-        Key::Numpad4 => 324,
-        Key::Numpad5 => 325,
-        Key::Numpad6 => 326,
-        Key::Numpad7 => 327,
-        Key::Numpad8 => 328,
-        Key::Numpad9 => 329,
-        //Key::NumpadDecimal => 330,
-        //Key::ContextMenu => 348,
-        Key::NumpadEquals => 336,
-        Key::F1 => 290,
-        Key::F2 => 291,
-        Key::F3 => 292,
-        Key::F4 => 293,
-        Key::F5 => 294,
-        Key::F6 => 295,
-        Key::F7 => 296,
-        Key::F8 => 297,
-        Key::F9 => 298,
-        Key::F10 => 299,
-        Key::F11 => 300,
-        Key::F12 => 301,
-        Key::F13 => 302,
-        Key::F14 => 303,
-        Key::F15 => 304,
-        Key::F16 => 305,
-        Key::F17 => 306,
-        Key::F18 => 307,
-        Key::F19 => 308,
-        Key::F20 => 309,
-        Key::F21 => 310,
-        Key::F22 => 311,
-        Key::F23 => 312,
-        _ => return None,
-    };
-    Some(code)
 }
